[Chore] Move `detokenizer_utils` to `vllm/tokenizers`
